gradio-resnet101-image-labels/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
brotli==1.2.0
certifi==2025.11.12
click==8.3.1
fastapi==0.127.1
ffmpy==1.0.0
filelock==3.20.1
fsspec==2025.12.0
gradio==6.2.0
gradio_client==2.0.2
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.21.0
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
torch==2.5.0+cpu
torchaudio==2.5.0+cpu
torchvision==0.20.0+cpu
python-magic==0.4.27